Upgrade to php >= 7.4, add strong type hints and remove useless PHPDoc

+++ Constraint/Constraints/GreaterThanOrEqual.php
@@ -8,13 +8,10 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
 final class GreaterThanOrEqual extends Constraint
 {
     /**
-     * @var int
+     * @var int|float|string
      */
     private $value;
 
-    /**
-     * @param array $options
-     */
     public function __construct(array $options = [])
     {
         parent::__construct($options);
@@ -22,30 +19,21 @@ final class GreaterThanOrEqual extends Constraint
         $this->value = $options['value'];
     }
 
-    /**
-     * @inheritdoc
-     */
     protected function getAttribute(): string
     {
         return 'data-parsley-gte';
     }
 
-    /**
-     * @inheritdoc
-     */
     protected function getValue(): string
     {
         return (string) $this->value;
     }
 
-    /**
-     * @inheritdoc
-     */
     protected function configureOptions(OptionsResolver $resolver): void
     {
         $resolver
             ->setRequired(['value'])
-            ->setAllowedTypes('value', ['int'])
+            ->setAllowedTypes('value', ['numeric'])
         ;
     }
 }
